+++ effective-gear/common/src/main/java/com/chimericdream/effectivegear/mixin/EG$PiglinAiMixin.java
@@ -4,6 +4,7 @@ import com.chimericdream.effectivegear.item.armor.TrimSetUtils;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.monster.piglin.PiglinAi;
 import net.minecraft.world.item.equipment.trim.TrimMaterials;
+import net.minecraft.world.item.equipment.trim.TrimPatterns;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -12,8 +13,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(PiglinAi.class)
 public class EG$PiglinAiMixin {
     @Inject(method = "isWearingSafeArmor", at = @At("HEAD"), cancellable = true)
-    private static void eg$treatFullGoldTrimAsSafeArmor(LivingEntity livingEntity, CallbackInfoReturnable<Boolean> cir) {
-        if (TrimSetUtils.isWearingFullTrim(livingEntity, TrimMaterials.GOLD)) {
+    private static void eg$treatFullGoldOrSnoutTrimAsSafeArmor(LivingEntity livingEntity, CallbackInfoReturnable<Boolean> cir) {
+        if (TrimSetUtils.isWearingFullTrim(livingEntity, TrimMaterials.GOLD) || TrimSetUtils.isWearingFullPattern(livingEntity, TrimPatterns.SNOUT)) {
             cir.setReturnValue(true);
         }
     }