+++ effective-gear/fabric/src/main/java/com/chimericdream/effectivegear/fabric/EffectiveGearFabric.java
@@ -3,10 +3,12 @@ package com.chimericdream.effectivegear.fabric;
 import net.fabricmc.api.ModInitializer;
 
 import com.chimericdream.effectivegear.EffectiveGearMod;
+import com.chimericdream.effectivegear.fabric.network.FabricServerNetworking;
 
 public final class EffectiveGearFabric implements ModInitializer {
     @Override
     public void onInitialize() {
         EffectiveGearMod.init();
+        FabricServerNetworking.init();
     }
 }