+++ effective-gear/fabric/src/main/java/com/chimericdream/effectivegear/fabric/client/EffectiveGearFabricClient.java
@@ -1,10 +1,20 @@
 package com.chimericdream.effectivegear.fabric.client;
 
+import com.chimericdream.effectivegear.client.Keybindings;
+import com.chimericdream.effectivegear.network.UseAbilityPayload;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 
 public final class EffectiveGearFabricClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
-        // This entrypoint is suitable for setting up client-specific logic, such as rendering.
+        Keybindings.init();
+
+        ClientTickEvents.END_CLIENT_TICK.register(client -> {
+            while (Keybindings.USE_ABILITY.consumeClick()) {
+                ClientPlayNetworking.send(UseAbilityPayload.INSTANCE);
+            }
+        });
     }
 }