+++ effective-gear/neoforge/src/main/java/com/chimericdream/effectivegear/neoforge/EffectiveGearNeoForge.java
@@ -4,10 +4,12 @@ import net.neoforged.fml.common.Mod;
 
 import com.chimericdream.effectivegear.EffectiveGearMod;
 import com.chimericdream.effectivegear.ModInfo;
+import com.chimericdream.effectivegear.neoforge.network.NeoForgeServerNetworking;
 
 @Mod(ModInfo.MOD_ID)
 public final class EffectiveGearNeoForge {
     public EffectiveGearNeoForge() {
         EffectiveGearMod.init();
+        NeoForgeServerNetworking.init();
     }
 }